Remove all channel in dope sheet editor and fix keyframe GUI to work with new McLib changes

+++ src/main/java/mchorse/aperture/client/gui/utils/GuiCameraEditorKeyframesEditor.java
@@ -42,7 +42,7 @@ public abstract class GuiCameraEditorKeyframesEditor<E extends GuiKeyframeElemen
     }
 
     @Override
-    public void setTick(long value)
+    public void setTick(double value)
     {
         super.setTick(value);
         this.editor.updateProfile();
